IMPALA-12589: Ignore FbFileDesc.absolutePath() if it is null/empty
IMPALA-11507 add absolute_path field to FbFileDesc in
CatalogObjects.fbs. This field is used as a fallback for the
relative_path field when the relative path of an Iceberg data file
cannot be obtained by the table location path. relative_path is never
set with null, while absolute_path may not be set and left as null.

IMPALA-11507 missed a case where both FileDescriptor.getRelativePath()
and FbFileDesc.absolutePath() are null/empty. This patch fixes the bug
so that FileDescriptor.getPath() and FileDescriptor.getAbsolutePath() do
not fall back to FbFileDesc.absolutePath() if it is null/empty. With
this fix, FileDescriptor.getPath() and FileDescriptor.getAbsolutePath()
will not return null.

Testing:
- Add test_scanner.py::TestSingleFileTable
- Pass core tests

Change-Id: I15d811430871cd1319d9560901318c1c3b89403c
Reviewed-on: http://gerrit.cloudera.org:8080/20769
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>