IMPALA-12629: Fix Iceberg V2 tests in non-HDFS environment
The UPDATE part 2 patch (https://gerrit.cloudera.org/#/c/20677/)
added new ranger tests that verifies that row filtering and column
masking are working on Iceberg V2 tables with delete files.

The Iceberg V2 tables we load during data loading are already written
and the position delete files have hard-coded file URIs that refer
to the data files. Unfortunately these URIs start with 'hdfs://...'.
Therefore these tables cannot be used in non-HDFS environment.

To quickly fix this, this patch introduces the 'IS_HDFS_ONLY' test
section. Tests that are annotated with it are only executed when
the underlying storage system is HDFS. This new test section can be
also used in the future to make it easier to write Impala/Hive interop
tests, because currently we can only run Hive in an HDFS environment.
So we could just annotate the tests that have HIVE_QUERY with
IS_HDFS_ONLY, and we could also add a HIVE_RESULTS section to check
the results Hive produces.

Testing:
 * Ran the tests with '-s' option in HDFS environment and verified
   that the IS_HDFS_ONLY tests are executed
 * Ran the ranger tests in Ozone environment and they passed

Change-Id: Ifa72041e281e3d5083adaee2f8a41bb1c2c8d19f
Reviewed-on: http://gerrit.cloudera.org:8080/20803
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>