IMPALA-12633: Remove DCHECK for slow SetQueryInflight
Removes the DCHECK that the original query is inflight before trying to
close it during a query retry. SetQueryInflight is a separate operation
the server performs after a query has started executing async, and it's
possible for the query to fail and retry before the server calls
SetQueryInflight. When that happens, we still need to perform cleanup
or the original request_state is never closed and we hit a different
DCHECK: "BlockOnWait() needs to be called!"

Adds an option to CloseClientRequestState for when we close a
ClientRequestState but the query is retrying with a new state. It
ensures that we bypass most of SetQueryInflight in case
CloseClientRequestState was called first.

Updates the message from DCHECK in ClientRequestState's destructor to
reflect that wait_thread_ is only reset in Finalize.

Adds a debug action and test where just the original query is delayed
during the SetQueryInflight call.

Change-Id: Ic17a5e12d9db61cb19306270174518a8dfd281a7
Reviewed-on: http://gerrit.cloudera.org:8080/20799
Reviewed-by: Riza Suminto <riza.suminto@cloudera.com>
Reviewed-by: Wenzhe Zhou <wzhou@cloudera.com>
Reviewed-by: Quanlong Huang <huangquanlong@gmail.com>
Tested-by: Michael Smith <michael.smith@cloudera.com>