IMPALA-12628: iceberg negative test fail in non-HDFS environments
There's a HIVE_QUERY section in iceberg-negative.test. Such sections
use Hive to execute statements. In non-HDFS environments like S3,
Ozone, etc. Hive is not configured properly, so we cannot use
it.

This patch adds the '---- IS_HDFS_ONLY' section to the relevant
test blocks (The HIVE_QUERY that creates the table, and the
Impala test that checks an error message).

Change-Id: I7987724f11d0c88d09577f52294ecc10d8ce39ce
Reviewed-on: http://gerrit.cloudera.org:8080/20812
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>
Tested-by: Zoltan Borok-Nagy <boroknagyz@cloudera.com>