IMPALA-11853: Fix formatted docs query options CSS
This is the release part of the CSS issue that was fixed by
IMPALA-11853.

The recent documentation formatting changes introduced the navigation
panel on the left. However, due to the length of the query options
navigation title these could overlap with the documentation paragraphs.

This commit removes the underscores from the navigation titles of the
query options, so browsers can break them into multiple lines.
Additionally, the "SET" and "Query Options for the SET Statement" pages
are merged to save some more space for the query option navigation
titles.

Change-Id: Ifcf12f401f3d4975c94ba78084ecf2c20267fad1
Reviewed-on: http://gerrit.cloudera.org:8080/20824
Tested-by: Tamas Mate <tmater@apache.org>
Reviewed-by: Tamas Mate <tmater@apache.org>