IMPALA-12661: Fix ASAN heap-use-after-free in IcebergMetadataScanNode
The ASAN builds detected that the IcebergMetadataScanNode uses heap
allocated memory after it has been freed.

In CreateFieldAccessors() method, during tree traversal, the
current_type variable is reassigned to its children which is part of
of the object. However, by the end of the assignment the rhs object will
be destroyed. To fix this issue, the variable was replaced with a pointer.

Testing:
 - Ran tests on ASAN build

Change-Id: I6df9c9cb6914a0c6c93b61aa0dd02acfdba68851
Reviewed-on: http://gerrit.cloudera.org:8080/20829
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>