IMPALA-12673: Table migration fails if partition contains '/'
Due to Iceberg #7612 migrating a table to Iceberg resulted in incorrect
data and stats if some of the string partition fields contained '/'
character. As a result we deliberately rejected migrating such tables.
Now that Impala uses an Iceberg version that has the fix we can allow
migrating such tables too.

Change-Id: I05b4ca44c7edb81cee6747f83a5bd82c5a4b5c44
Reviewed-on: http://gerrit.cloudera.org:8080/20845
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>