IMPALA-12683: Fix wrong event time for batched events
When updating the last-synced-event-id after processing a batch of
partition events, we use the last event id. We should do the same when
updating last-synced-event-time. However, currently BatchPartitionEvent
uses getEventTime() from the parent class. It actually returns the event
time of the first event. We should override it to use the last event.

Tests
 - Ran MetastoreEventsProcessorTest.testDisableEventSyncFlag 200 times.

Change-Id: I82efe18dd28fe8af47f8c66cc8c5eb8e6f8dfd2b
Reviewed-on: http://gerrit.cloudera.org:8080/20864
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>