IMPALA-12688: Support JSON profile imports in webUI
A button has been added to import query profiles on the queries page.

The button opens a system dialog to browse only "*.json" files from
client's local file system. This allows multiple JSON query profiles
to be imported.

JSON profiles are read from file system using FileReader API and then
these imported query profiles are stored in a NoSQL database using
browser's IndexedDB API.

If an error is encountered while parsing JSON profile, it is displayed
on the queries page.

IndexedDB is a large-scale NoSQL database system, that can be read
and written to asynchronously. Hence, the file size and number of JSON
profiles is only limited by the storage available on the client system.

The storage limit is browser dependent, in chromium based browsers
upto 80% of local storage may be used by IndexedDB.

The profile is parsed for the following attributes, which will be
displayed on the queries page. Number of imported queries is also
displayed.

- Query ID
- User
- Default Db
- Query Type
- Start Time
- End Time
- Bytes Read
- Bytes Sent
- Query State
- No. of rows fetched
- Resource Pool
- Query Statement

After importing or deleting query profiles, page will scroll back
to the imported query profiles section.

The queries are sorted based on descending order of the "Start Time".

A button is displayed beside the import button to easily clear all
the imported queries at once.

On opening any of the imported queries, its query timeline will be
displayed. Currently the following tabs are supported for imported
queries.

- Query Statement
- Query Timeline
- Query Text Plan

Syntax highlighting on query statement tab has also been supported.

Change-Id: Ife6eb59bf2030fd19fc92aaf134eb51c609e04d0
Reviewed-on: http://gerrit.cloudera.org:8080/20867
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>