IMPALA-12690: Add TpcdsCpuCostPlannerTest
This patch adds TpcdsCpuCostPlannerTest, which is a copy of
TpcdsPlannerTest with COMPUTE_PROCESSING_COST option enabled.
PROCESSING_COST_MIN_THREADS and MAX_FRAGMENT_INSTANCES_PER_NODE are set
to 2 and 16 accordingly to match the same options in
PlannerTest#testProcessingCost. PlannerTest#testProcessingCost is
reduced to contain only corner case queries that is not subset of
TPC-DS. TpcdsCpuCostPlannerTest runs for around 30 second.

Testing:
- Pass FE tests.

Change-Id: If04fa584db5f13db0dd656ec9d99f7204c05d75d
Reviewed-on: http://gerrit.cloudera.org:8080/20872
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>