IMPALA-12582: Fix crash when enabling MIN_MAX RuntimeFilter in Nested Loop Join
This patch fixes an issue reported by IMPALA-12582, where enabling
MIN_MAX RuntimeFilter for a specific query would cause the executor to
crash. The direct cause of the crash was an out-of-bounds access to
input_vals in the ScalarFnCall::InterpretEval() function, but the root
cause was actually due to the related ScalarExprEvaluator not invoking
the Open() function.

Testing:
 - Added new E2E test case about this issue.

Change-Id: Iba951796d52f109c419587c444840adbb2d44f5d
Reviewed-on: http://gerrit.cloudera.org:8080/20891
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>
Tested-by: Csaba Ringhofer <csringhofer@cloudera.com>