IMPALA-12448: Avoid getting stuck when refreshing a non-existent partition
In some cases, we shouldn't wait for the catalog version of 'result'
because the table version has not changed after resetting metadata,
such as refreshing a non-existent partition. If not, the query may be
stuck for a long time. If the catalog version of 'result' is less
than or equal to 'oldestTopicUpdateToGc_', return the catalog version
of 'result' directly.

Tests:
* added e2e tests

Change-Id: Iace7cdadda300b03896f92415822266354421887
Reviewed-on: http://gerrit.cloudera.org:8080/20490
Reviewed-by: Quanlong Huang <huangquanlong@gmail.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>